Use "int64_t" instead of "long long" in lib/algebra

+++ aoj/ALDS1/ALDS1B/main-bingcd.cpp
@@ -1,10 +1,10 @@
 #include <bits/stdc++.h>
 
-inline unsigned long long gcd(unsigned long long a, unsigned long long b) {
+inline uint64_t gcd(uint64_t a, uint64_t b) {
     if (a == 0) return b;
     if (b == 0) return a;
 
-    unsigned long long g = 1;
+    uint64_t g = 1;
     while (b > 0) {
         if ((a & 1) == 0 && (b & 1) == 0) {
             g <<= 1;
